feature: 왓치 리스트 새로고침 추가

+++ app/src/main/java/com/hong7/coinnews/ui/feature/watchlist/search/SearchWatchListViewModel.kt
@@ -55,7 +55,12 @@ class SearchWatchListViewModel @Inject constructor(
     fun toggleSelected(coin: Coin) {
         viewModelScope.launch {
             if (watchListCoins.value.contains(coin.id)) {
-                watchListRepository.removeWatchListCoin(coin)
+                if (watchListCoins.value.size >= 10) {
+                    // TODO
+                } else {
+                    watchListRepository.removeWatchListCoin(coin)
+                }
+
             } else {
                 watchListRepository.addWatchListCoin(coin)
             }
